+++ src/rex.deploy/NEWS.rst
@@ -10,6 +10,7 @@
 
 * Added a workaround for PostgreSQL dropping constraint comments.
 * Allow deleted fields in ``with`` clause.
+* Fixed incompatibility with ``tweak.override`` HTSQL plugin.
 
 
 2.3.1 (2015-04-07)